+++ src/extension.ts
@@ -1,21 +1,22 @@
+import { StateBag } from './stateBag';
 import * as vscode from "vscode";
 import * as configuration from './configuration';
 import * as runner from './runner';
 import * as contextHelpers from "./contextHelpers";
 
 export function activate(context: vscode.ExtensionContext) {
+	const state = {} as StateBag;
 
 	let disposables = [];
 	disposables.push(vscode.commands.registerCommand('extension.runChutzpah', (uri: vscode.Uri) => {	
-		runChutzpah(uri,false,false);
+		runChutzpah(uri,false,false,state);
 	}));
 	disposables.push(vscode.commands.registerCommand('extension.runChutzpahInBrowser', (uri: vscode.Uri) => {
-		runChutzpah(uri,true,false);
+		runChutzpah(uri,true,false,state);
 	}));
 	disposables.push(vscode.commands.registerCommand('extension.runChutzpahWithCoverage', (uri: vscode.Uri) => {
-		runChutzpah(uri,true,true);
+		runChutzpah(uri,true,true,state);
 	}));
-
 	context.subscriptions.push(...disposables);
 }
 
@@ -26,22 +27,22 @@ export function deactivate() {}
  * @param uri File or folder path to test
  * @param openBrowser Open in browser for debug
  */
-export function runChutzpah(uri: vscode.Uri, openBrowser: boolean, coverage: boolean): boolean {
+export function runChutzpah(uri: vscode.Uri, openBrowser: boolean, coverage: boolean, state: StateBag): boolean {
 
 	var chutzpahPath = configuration.getChutzpahPath();
-	if (chutzpahPath == "") {
+	if (chutzpahPath === "") {
 		vscode.window.showErrorMessage("Chutzpah Path is not valid");
 		return false;
 	}
 	var parallelism = configuration.getParallelism();
 	var testPath = contextHelpers.getPathFromUri(uri);
 
-	let args = [testPath,"/engine","chrome"];
+	let args = [`"${testPath}"`,"/engine","chrome"];
 
 	let disableCORS = configuration.getDisableCORS();
 	if (disableCORS) {
 		args.push("/browserArgs");
-		args.push(`"--disable-web-security --user-data-dir=${contextHelpers.getChromeProfilePath()}"`);
+		args.push(`--disable-web-security --user-data-dir="${contextHelpers.getChromeProfilePath()}"`);
 	}
 
 	if (openBrowser)
@@ -49,10 +50,10 @@ export function runChutzpah(uri: vscode.Uri, openBrowser: boolean, coverage: boo
 	if (parallelism)
 		args.push(...["/parallelism",parallelism.toString()]);
 	if (coverage)
-		args.push("/coverage")
+		args.push("/coverage");
 
 	if (!openBrowser) {
-		runner.spawnTests(chutzpahPath,args,testPath);
+		state.outputChannel = runner.spawnTests(chutzpahPath,args,testPath,state.outputChannel);
 	} else {
 		var terminal = runner.selectTerminal();
 		runner.terminalTests(chutzpahPath,args,terminal);